Add support for using a KeyName element within the KeyInfo block.
The KeyName element contains a string value (in which white space is
significant) which may be used by the signer to communicate a key
identifier to the recipient. Typically, KeyName contains an identifier
related to the key pair used to sign the message, but it may contain
other protocol-related information that indirectly identifies a key
pair. (Common uses of KeyName include simple string names for keys, a
key index, a distinguished name (DN), an email address, etc.)