verify: Add tests for XAdES-X-L form
The XAdES-X-L form includes the full certificate path up to the root
certificate, so that the verifier can accept any signature from a
certificate that has some path to a trusted root. This is described at:

    https://www.w3.org/TR/XAdES/#Syntax_forXAdES-X-L_form

This commit adds three certificates (and a script to re-generate them
if necessary): root, intermediate, and leaf. In the XAdES-X-L test case,
we sign with the leaf certificate but include all three in the
certificate chain. Verifying these works when the root certificate is
given as a trusted root, but fails when it is not considered a trusted
root.