Add options to exclude the C14N Transform element in signatures (#274)
This patch adds a flag 'exclude_c14n_transform_element' (boolean,
defaults to False) to XMLSigner.sign().

If set to True, the transform is still applied, but it is not added to
the Transforms element in ds:SignedInfo/ds:Reference.

Conversely, this patch adds an option to XMLVerifier.verify() to specify
a default canonicalization algorithm in the case where SignedInfo does
not include the c14n Transform information.

Rationale:

In most cases, the Transforms element should indeed contain every
modification made to the source data before creating or verifying the
signature, and indeed, the necessity to add a default algorithm in
verify() shows why. However, there are specifications that make use of
XML Signatures, where the canonicalization algorithm is fixed, and which
even go so far as to forbid naming the c14n algorithm in the Transforms
element.

This patch aims to make it possible to support those specifications.